Change the track command file function
Replace the dash symbol with current user in created by field
